Minor code improvements #2 Forgot to save files before previous commit

+++ Project3/ValidationFunctions.cpp
@@ -388,7 +388,7 @@ void ValidationCheck::LanguagePacks(Validation::Status& status)
 	std::vector<std::string> lang_files = MeridorConsoleLib::GetFilesInDirectory(FolderName::language);
 	bool good_files = false;
 
-	for (int i = 0; i < lang_files.size(); ++i)
+	for (size_t i = 0; i < lang_files.size(); ++i)
 	{
 		FileManagement::LoadLanguagePack(FolderName::language + '\\' + lang_files[i]);
 		if (
